Add warning about labels in direct page
If you have a label to a variable in the direct page and use that label
as an address for an instruction, we can't automatically figure out that
that is the case without doing multiple compilation & label resolution
attempts (which would make the assembly process undecidable in some
situations, as far as I can tell). Therefore, labels are always treated
as full addresses (16 bit) if applicable. This warning will thell the
user about it + the possibility of specifying a label in the direct page
explicitly in the future (see #1). The warning is only emitted if anyone
uses the label in an address position.