Improve invalid number errors
By not checking the radix validity early, we can initially accept a
generic number-like string, build a token and then report Rust-internal
parsing errors with proper error location. This makes number errors much
better in the normal case, which is radix mismatch such as hex digits in
a decimal number.