Remove fixed
Fixed is a nice crate but we don't need it; the SPC700 *doesn't* use
fixed-point arithmetic but simulates it through multiplication and
shifting. With this change, we probably don't emulate hardware more
closely (the fixed-point emulation is equivalent and already perfect),
but we improve speed to 1.6MB/s from 1.3MB/s.