Respect reference calculations and resulting offsets
This makes all direct page-ness calculations based on the actual address
obtained, not whether a reference used in the calculation is in the
direct page or not. Therefore, we can remove collecting references in
and out of the direct page (we need to consider all references and their
current "proposed" addresses).

Fixes #38