Fix multi-label locals
The labels are now wired up in a way that the last global label is the
actual label of the instruction in question, so that all the "current
global label" code is able to directly detect that this global label is
the global label for all the code following it. This is how a human
expects multi-labels to work, though it's nontrivial to implement.