Don't force-prefix references in numbers with "$"
These would previously stack a lot, with no easily discernible
information being added that way. Since it's not important to see
immediately how a resolved value was derived from references (rather,
it's important to see what the value is when the source code is
available), we can drop this information.